cys edit Pong

+++ wingchun/td/TDEngineHuobi.cpp
@@ -214,13 +214,13 @@ void TDEngineHuobi::Ping(struct lws* conn)
     strncpy((char *)msg+LWS_PRE, strPing.c_str(), length);
     int ret = lws_write(conn, &msg[LWS_PRE], length,LWS_WRITE_TEXT);
 }
-void TDEngineHuobi::Pong(struct lws* conn,int ping){
+void TDEngineHuobi::Pong(struct lws* conn,long long ping){
     KF_LOG_INFO(logger,"[Pong] pong the ping of websocket");
     StringBuffer sbPing;
     Writer<StringBuffer> writer(sbPing);
     writer.StartObject();
     writer.Key("pong");
-    writer.Int(ping);
+    writer.Int64(ping);
     writer.EndObject();
     std::string strPong = sbPing.GetString();
     unsigned char msg[512];
@@ -263,7 +263,7 @@ void TDEngineHuobi::on_lws_data(struct lws* conn, const char* data, size_t len)
                     on_lws_receive_orders(conn,json);
                 }
             } else if (op == "ping") {
-                int ping=json["ts"].GetInt();
+                long long ping=json["ts"].GetInt64();
                 Pong(conn,ping);
             } else if (op == "auth") {
                 isAuth=huobi_auth;
@@ -273,7 +273,7 @@ void TDEngineHuobi::on_lws_data(struct lws* conn, const char* data, size_t len)
         } else if (json.HasMember("ch")) {
 
         } else if (json.HasMember("ping")) {
-            int ping=json["ts"].GetInt();
+            long long ping=json["ts"].GetInt64();
             Pong(conn,ping);
         } else if (json.HasMember("subbed")) {
